assets/promptflow/ui-evaluations-flow/requirements.txt
azureml-metrics[generative-ai]
promptflow
promptflow-evals
azure-identity